refactor structure of pt component
