+++ benchmark/src/components/pytorch_image_classifier/torch_helper/image_io.py
