+++ benchmark/src/components/pytorch_image_classifier/torch_helper/model/__init__.py
