+++ benchmark/src/components/pytorch_image_classifier/torch_helper/model/model_loader.py
