+++ benchmark/src/components/pytorch_image_classifier/torch_helper/model/swin_models.py
