+++ benchmark/src/components/pytorch_image_classifier/torch_helper/model/test_model.py
