+++ benchmark/src/components/pytorch_image_classifier/torch_helper/model/torchvision_models.py
